# server/internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"time"

	"github.com/google/uuid"
)

type ClientError struct {
	ID         int64
	UserID     *uuid.UUID
	RequestID  *string
	Message    string
	Stack      *string
	Route      *string
	ClientKind string
	AppVersion string
	OccurredAt time.Time
	ReceivedAt time.Time
}

type Identity struct {
	Provider  string
	Subject   string
	UserID    uuid.UUID
	CreatedAt time.Time
}

type Session struct {
	TokenHash         []byte
	UserID            uuid.UUID
	ClientKind        string
	CreatedAt         time.Time
	LastSeenAt        time.Time
	IdleExpiresAt     time.Time
	AbsoluteExpiresAt time.Time
}

type User struct {
	ID        uuid.UUID
	Name      string
	Role      string
	Locale    string
	CreatedAt time.Time
}